Исправил замечания по предыдущему ПР, добавил компонент для пустого списка точек
