+++ src/utils.js
@@ -48,3 +48,11 @@ export const createElement = (template) => {
   return newElement.firstChild;
 };
 
+export const ifEscKeyPressed = (evt) => {
+  if (evt.key === `Escape` || evt.key === `Esc`) {
+    return true;
+  }
+  return false;
+};
+
+
