+++ src/view/empty/empty.tpl.js
@@ -1,0 +1,3 @@
+export const createEmptyTemplate = () => {
+  return `<p class="trip-events__msg">Click New Event to create your first point</p>`;
+};
