+++ src/view/filters/filters.tpl.js
@@ -1,6 +1,4 @@
-import {createElement} from "../utils.js";
-
-const createFiltersTemplate = (filters) => {
+export const createFiltersTemplate = (filters) => {
 
   const filterItems = filters.map((value) => {
     const valueLowerCase = value.toLowerCase();
@@ -15,22 +13,3 @@ const createFiltersTemplate = (filters) => {
   <button class="visually-hidden" type="submit">Accept filter</button>
 </form>`;
 };
-
-export default class FiltersView {
-  constructor(data) {
-    this._element = null;
-    this._data = data;
-  }
-  getTemplate() {
-    return createFiltersTemplate(this._data);
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
-  }
-}
