+++ src/view/form-list/form-list.tpl.js
@@ -1,0 +1,4 @@
+export const createFormListTemplate = () => {
+  return `<ul class="trip-events__list">
+  </ul>`;
+};
